Commit 4 weeks later, I mostly remember about doubting the way I was trying to organize the rooms
I wanted to create objects with method the player can interact directly
with, somekind of direct control but through text

Now I think a room is a list of regex&action pairs.
The regex is used to validate if the action should be ran
Actions are behaviour (methods) to interact with the world that receive the player, room state and extracted text.
The player could have a reference to the room he's currently in for easy
access, but having a room id would be great for logging and
serialization.